requirements.txt
git+https://github.com/Altruix/pyrogram
tgcrypto
aiofiles
pyromod
python-dotenv
psycopg2-binary==2.8.6
sqlalchemy==1.3.23
